requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Data validation
pydantic==2.7.1

# Networking / URL parsing
tldextract==5.1.2
dnspython==2.6.1
requests==2.32.3

# WHOIS
python-whois==0.8.0

# SSL / Crypto
cryptography==42.0.7

# Async MongoDB (for next step)
motor==3.5.1

# Environment management
python-dotenv==1.0.1

# Logging & utilities
loguru==0.7.2

scikit-learn==1.5.0
pandas==2.2.2
joblib==1.4.2